address jridgewell comments
